[FE-1266-2] moved Charts to "src/components/reportBuilder"
