+++ src/components/reportBuilder/index.js
@@ -1,2 +1,3 @@
 export { default as AggregatedMetrics } from './AggregatedMetrics';
 export { default as CompareByAggregatedMetrics } from './CompareByAggregatedMetrics';
+export { ChartGroups, Charts } from './Charts';
