+++ src/pages/reportBuilder/components/Charts.js
@@ -1,172 +1,8 @@
-import React, { useMemo, useState } from 'react';
-import _ from 'lodash';
-import { getLineChartFormatters } from 'src/helpers/chart';
-import LineChart from './LineChart';
-import METRICS_UNIT_CONFIG from 'src/config/metrics-units';
-import { Box, Stack, Panel } from 'src/components/matchbox';
-import { useSparkPostQuery } from 'src/hooks';
-import { getTimeSeriesDeliverabilityMetrics } from 'src/helpers/api';
-import {
-  getMetricsFromKeys,
-  getQueryFromOptionsV2 as getQueryFromOptions,
-  transformData,
-} from 'src/helpers/metrics';
-import { REPORT_BUILDER_FILTER_KEY_MAP } from 'src/constants';
+import React from 'react';
 import { useReportBuilderContext } from '../context/ReportBuilderContext';
-import { ApiErrorBanner } from 'src/components';
-import Loading from 'src/components/loading/PanelLoading';
-import { Heading } from 'src/components/text';
-const DEFAULT_UNIT = 'number';
-
-function getUniqueUnits(metrics) {
-  return _.uniq(metrics.map(({ unit = DEFAULT_UNIT }) => unit));
-}
+import { ChartGroups } from 'src/components/reportBuilder';
 
 export default function ChartContainer() {
   const { state: reportOptions } = useReportBuilderContext();
   return <ChartGroups reportOptions={reportOptions} />;
-}
-
-export function ChartGroups(props) {
-  const { reportOptions } = props;
-  const { comparisons } = reportOptions;
-  const hasComparisons = Boolean(comparisons.length);
-  const [activeChart, setActiveChart] = useState(null);
-
-  if (!hasComparisons) {
-    return (
-      <Panel.Section>
-        <Charts
-          activeChart={activeChart}
-          setActiveChart={setActiveChart}
-          id="chart"
-          reportOptions={reportOptions}
-        />
-      </Panel.Section>
-    );
-  }
-
-  return (
-    <>
-      {comparisons.map((compareFilter, index) => {
-        const filterType = REPORT_BUILDER_FILTER_KEY_MAP[compareFilter.type];
-
-        // Appends each compared filter as a new filter for individual requests
-        const comparedFilters = [
-          ...reportOptions.filters,
-          { AND: { [filterType]: { eq: [compareFilter] } } },
-        ];
-        return (
-          <Panel.Section key={`chart_group_${index}`}>
-            <Stack>
-              <Box>
-                <Heading looksLike="h5" as="h3">
-                  {compareFilter.value}
-                </Heading>
-              </Box>
-              <Box>
-                <Charts
-                  activeChart={activeChart}
-                  setActiveChart={setActiveChart}
-                  id={`chart_group_${index}`}
-                  reportOptions={{ ...reportOptions, filters: comparedFilters }}
-                />
-              </Box>{' '}
-            </Stack>
-          </Panel.Section>
-        );
-      })}
-    </>
-  );
-}
-
-export function Charts(props) {
-  const { reportOptions, activeChart, setActiveChart, id } = props;
-  const { comparisons, metrics } = reportOptions;
-
-  // Prepares params for request
-  const formattedMetrics = useMemo(() => {
-    return getMetricsFromKeys(metrics, true);
-  }, [metrics]);
-  const formattedOptions = useMemo(() => {
-    return getQueryFromOptions({ ...reportOptions, metrics: formattedMetrics });
-  }, [reportOptions, formattedMetrics]);
-  const { precision, to } = formattedOptions;
-
-  // API request
-  const { data: rawChartData, status: chartStatus, refetch: refetchChart } = useSparkPostQuery(
-    () => {
-      return getTimeSeriesDeliverabilityMetrics(formattedOptions);
-    },
-    {
-      refetchOnWindowFocus: false,
-    },
-  );
-
-  const chartData = useMemo(() => {
-    return transformData(rawChartData, formattedMetrics);
-  }, [rawChartData, formattedMetrics]);
-
-  const formatters = getLineChartFormatters(precision, to);
-  //Separates the metrics into their appropriate charts
-  const charts = getUniqueUnits(formattedMetrics).map(unit => ({
-    metrics: formattedMetrics.filter(metric => metric.unit === unit),
-    ...METRICS_UNIT_CONFIG[unit],
-  }));
-  let height = 150;
-
-  switch (charts.length * (comparisons.length || 1)) {
-    case 1:
-      height = 400;
-      break;
-    case 2:
-      height = 200;
-      break;
-    default:
-      break;
-  }
-
-  if (chartStatus === 'loading' || chartStatus === 'idle') {
-    return <Loading as={Box} minHeight="200px" />;
-  }
-
-  if (chartStatus === 'error') {
-    return (
-      <ApiErrorBanner
-        reload={refetchChart}
-        status="muted"
-        title="Unable to load report"
-        message="Please try again"
-      />
-    );
-  }
-
-  return (
-    <Box>
-      <Stack>
-        {charts.map((chart, index) => (
-          <Box key={`chart-${index}`} onMouseOver={() => setActiveChart(`${id}_chart_${index}`)}>
-            <LineChart
-              height={height}
-              syncId="summaryChart"
-              data={chartData}
-              precision={precision}
-              showTooltip={activeChart === `${id}_chart_${index}`}
-              lines={chart.metrics.map(({ name, label, stroke }) => ({
-                key: name,
-                dataKey: name,
-                name: label,
-                stroke,
-              }))}
-              {...formatters}
-              yTickFormatter={chart.yAxisFormatter}
-              yLabel={chart.label}
-              tooltipValueFormatter={chart.yAxisFormatter}
-              showXAxis={index === charts.length - 1}
-            />
-          </Box>
-        ))}
-      </Stack>
-    </Box>
-  );
 }
